TSNE fixes/replacements p.1
